+++ backend/main.go
@@ -11,6 +11,10 @@ import (
 	"github.com/rs/cors"
 	"gitlab.msu.edu/team-corewell-2025/routes/llm"
 	"gitlab.msu.edu/team-corewell-2025/routes/supabase"
+	patients "gitlab.msu.edu/team-corewell-2025/routes/supabase/patients"
+	prescriptions "gitlab.msu.edu/team-corewell-2025/routes/supabase/prescriptions"
+	results "gitlab.msu.edu/team-corewell-2025/routes/supabase/results"
+	tasks "gitlab.msu.edu/team-corewell-2025/routes/supabase/tasks"
 )
 
 func main() {
@@ -24,7 +28,13 @@ func main() {
 	// Create client
 	url := os.Getenv("SUPABASE_URL")
 	key := os.Getenv("SUPABASE_KEY")
-	supabase.InitClient(url, key)
+	supa := supabase.InitClient(url, key)
+
+	// Dependency Injection
+	var ph patients.PatientService = &patients.PatientHandler{Supabase: supa}
+	var prh prescriptions.PrescriptionService = &prescriptions.PrescriptionHandler{Supabase: supa}
+	var rh results.ResultService = &results.ResultHandler{Supabase: supa}
+	var th tasks.TaskService = &tasks.TaskHandler{Supabase: supa}
 
 	// API Gateway
 	m := mux.NewRouter()
@@ -33,7 +43,7 @@ func main() {
 	c := cron.New(cron.WithSeconds())
 	_, err = c.AddFunc("@daily", func() {
 		fmt.Println("Job Scheduler Triggered. ")
-		err = supabase.GenerateTasks(3, 3, 3, false)
+		err = th.GenerateTasks(3, 3, 3, false)
 		if err != nil {
 			fmt.Println("Error in cron job:", err)
 		}
@@ -45,7 +55,6 @@ func main() {
 	}
 	c.Start()
 
-	// Endpoints
 
 	// Auth
 	m.HandleFunc("/addUser", supabase.SignUpUser).Methods("POST")
@@ -55,17 +64,22 @@ func main() {
 
 	// Patients
 	patientsRouter := m.PathPrefix("/patients").Subrouter()
-	patientsRouter.HandleFunc("", supabase.GetPatients).Methods("GET")
-	patientsRouter.HandleFunc("/{id}", supabase.GetPatientByID).Methods("GET")
-	patientsRouter.HandleFunc("/{id}/prescriptions", supabase.GetPrescriptionsByPatientID).Methods("GET")
-	patientsRouter.HandleFunc("/{id}/results", supabase.GetResultsByPatientID).Methods("GET")
-	// UPDATED: Change method from GET to POST to support the giga json payload.
+	patientsRouter.HandleFunc("", ph.GetPatients).Methods("GET")
+	patientsRouter.HandleFunc("/{id}", ph.GetPatientByID).Methods("GET")
+	patientsRouter.HandleFunc("/{id}/prescriptions", prh.GetPrescriptionsByPatientID).Methods("GET")
+	patientsRouter.HandleFunc("/{id}/results", rh.GetResultsByPatientID).Methods("GET")
 	patientsRouter.HandleFunc("/{id}/llm-response", llm.PostLLMResponseForPatient).Methods("POST")
+
+	// Flagging Feature
+	m.HandleFunc("/flaggedPatients", ph.GetFlaggedPatients).Methods("GET")
+	m.HandleFunc("/addFlag", ph.AddFlaggedPatient).Methods("POST")
+	m.HandleFunc("/removePatient", ph.RemoveFlaggedPatient).Methods("POST")
+	m.HandleFunc("/keepPatient", ph.KeepPatient).Methods("POST")
 
 	// Prescriptions
 	prescriptionsRouter := m.PathPrefix("/prescriptions").Subrouter()
-	prescriptionsRouter.HandleFunc("", supabase.GetPrescriptions).Methods("GET")
-	prescriptionsRouter.HandleFunc("/{id}", supabase.GetPrescriptionByID).Methods("GET")
+	prescriptionsRouter.HandleFunc("", prh.GetPrescriptions).Methods("GET")
+	prescriptionsRouter.HandleFunc("/{id}", prh.GetPrescriptionByID).Methods("GET")
 
 	// Students
 	studentsRouter := m.PathPrefix("/students").Subrouter()
@@ -74,30 +88,21 @@ func main() {
 
 	// Results
 	resultsRouter := m.PathPrefix("/results").Subrouter()
-	resultsRouter.HandleFunc("", supabase.GetResults).Methods("GET")
-	resultsRouter.HandleFunc("/{id}", supabase.GetResultByID).Methods("GET")
-
-	// Flagging Feature
-	m.HandleFunc("/flaggedPatients", supabase.GetFlaggedPatients).Methods("GET")
-	m.HandleFunc("/addFlag", supabase.AddFlaggedPatient).Methods("POST")
-	m.HandleFunc("/removePatient", supabase.RemoveFlaggedPatient).Methods("POST")
-	m.HandleFunc("/keepPatient", supabase.KeepPatient).Methods("POST")
+	resultsRouter.HandleFunc("", rh.GetResults).Methods("GET")
+	resultsRouter.HandleFunc("/{id}", rh.GetResultByID).Methods("GET")
 
 	// Endpoints for tasks (generating, getting, completing, etc.)
-	m.HandleFunc("/generateTasks", supabase.GenerateTasksHTMLWrapper).Methods("POST")
-	m.HandleFunc("/{student_id}/tasks", supabase.GetTasksByStudentID).Methods("POST", "OPTIONS") //had to make this post bc the function expects a body
+	m.HandleFunc("/generateTasks", th.GenerateTasksHTMLWrapper).Methods("POST")
+	m.HandleFunc("/{student_id}/tasks", th.GetTasksByStudentID).Methods("POST", "OPTIONS") //had to make this post bc the function expects a body
 	//hardcoded body to show incomplete tasks ^^^
-	m.HandleFunc("/{student_id}/tasks/week", supabase.GetTasksByWeekAndDay).Methods("GET")
-	m.HandleFunc("/{student_id}/tasks/{task_id}", supabase.GetTaskByID).Methods("GET")
-	m.HandleFunc("/{student_id}/tasks/{task_id}/completeTask", supabase.CompleteTask).Methods("POST")
+	m.HandleFunc("/{student_id}/tasks/week", th.GetTasksByWeekAndDay).Methods("GET")
+	m.HandleFunc("/{student_id}/tasks/{task_id}", th.GetTaskByID).Methods("GET")
+	m.HandleFunc("/{student_id}/tasks/{task_id}/completeTask", th.CompleteTask).Methods("POST")
 
 	//Student Instructor Assignment Feature
 	m.HandleFunc("/instructors", supabase.GetInstructors).Methods("GET")
 	m.HandleFunc("/instructors/{id}/students", supabase.GetInstructorStudents).Methods("GET")
 	m.HandleFunc("/addStudent", supabase.AddStudentToInstructor).Methods("POST")
-
-	// Misc
-	m.HandleFunc("/messageRequest", llm.RequestMessage).Methods("POST")
 
 	// Allow API requests from frontend
 	handler := cors.New(cors.Options{
