+++ backend/routes/supabase/results/interfaces.go
@@ -1,0 +1,9 @@
+package results
+
+import "net/http"
+
+type ResultService interface {
+	GetResults(w http.ResponseWriter, r *http.Request)
+	GetResultByID(w http.ResponseWriter, r *http.Request)
+	GetResultsByPatientID(w http.ResponseWriter, r *http.Request)
+}
